access to database from subject
- Load subjects from database
- Load subjects from XML

+++ app/src/main/java/es/uniovi/university_management/model/TeacherSubjectEntity.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class TeacherSubjectEntity(
-    var teacherId: Int,
-    var subjectId: Int
+    var teacherId: Long,
+    var subjectId: Long
 ) {
     @PrimaryKey(autoGenerate = true)
     var id: Int? = null
